add _form.html and edit/create update

+++ src/Controller/PinsController.php
@@ -76,10 +76,14 @@ class PinsController extends AbstractController
     #[Route('/pins/{id<[0-9]+>}/delete', name: 'app_pins_delete', methods: ['DELETE'])]
     public function delete(
         Pin $pin,
+        Request $request,
         EntityManagerInterface $em): Response
     {
-        $em->remove($pin);
-        $em->flush();
+        if ($this->isCsrfTokenValid('pin_deletion_' . $pin->getId(), $request->request->get('csrf_token')))
+        {
+            $em->remove($pin);
+            $em->flush();
+        }
 
         return $this->redirectToRoute('app_home');
     }
